requirements.txt
streamlit
numpy
pillow
tensorflow
opencv-python
